Cast labels to float tensor in data.py

+++ deeplearning/exercise-4/src_to_implement/data.py
@@ -40,7 +40,7 @@ class ChallengeDataset(Dataset):
         if label == 1:
             label = torch.tensor([0, 1])
         else:
-            label = torch.tensor([1, 0])
+            label = torch.tensor([1, 0], dtype=torch.float64)
         return image, label
